+++ backend/src/main/java/com/learningplatform/learning_platform/service/UserService.java
@@ -6,15 +6,12 @@ import org.springframework.http.ResponseEntity;
 import java.util.List;
 
 /**
- * Service interface for User management
- * Authors:
- * - Rasika
- * Date: 8/9/2024
+ * Service interface for User management.
  */
 public interface UserService {
     List<User> getAllUsers();
     User createUser(User user);
-    ResponseEntity<User> getUserById(Long id);
-    ResponseEntity<User> updateUser(Long id, User user);
-    ResponseEntity<Void> deleteUser(Long id);
+    ResponseEntity<User> getUserById(Integer id);
+    ResponseEntity<User> updateUser(Integer id, User user);
+    ResponseEntity<Void> deleteUser(Integer id);
 }
